lib/engine/job_service/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
apscheduler
supabase
httpx
